+++ contrib/opencensus-ext-flask/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add exception information to span attributes
+([#1188](https://github.com/census-instrumentation/opencensus-python/pull/1188))
+
 ## 0.8.1
 Released 2022-08-03